🎨 Wrap REST API with GraphQL
## Todo
It seems that when value bigger than 9.0 is requested as `rating` param, the REST API does not response the correct JSON array. It returns a list of movies of lower ratings between 4~5, but I can not find the reason of this. It might just be the problem of yts api, perhaps.